Use buildkit for building and keeping apt cache
buildkit will package the apt cache along with the image in the same repository.

+++ installer/stretch/lib/functions.sh
@@ -83,9 +83,12 @@ function clean()
     done
 
     apt-get auto-remove -qq -y
-    apt-get clean
 
-    rm -rf /var/lib/apt/lists/*
+    if [ "${CLEAN_SKIP_APT_CACHE:-}" != true ]; then
+        apt-get clean
+
+        rm -rf /var/lib/apt/lists/*
+    fi
 }
 
 function install()
